requirements.txt
numpy~=1.21.6
imageio~=2.19.2
pi~=0.1.2
opencv-python~=4.5.5.64
pandas~=1.3.5
scikit-image~=0.16.2
scipy~=1.7.3
pythran~=0.11.0
Pillow~=8.4.0
pip~=22.1
wheel~=0.37.1
gast~=0.5.3
pytz~=2022.1
matplotlib~=3.5.2
Cython~=0.29.30
setuptools~=62.3.2
openpyxl~=3.0.9
python-dateutil~=2.8.2
xlrd~=1.2.0
pyserial~=3.5
beniget~=0.4.1
tifffile~=2021.11.2
ply~=3.11
networkx~=2.6.3
six~=1.16.0
fonttools~=4.33.3
pyparsing~=3.0.9
cycler~=0.11.0
packaging~=21.3
kiwisolver~=1.4.2
easygui~=0.98.3